pyproject.toml
[project]
name = "django-saas-project"
version = "0.1.0"
description = ""
authors = [
    {name = "vibbsdod",email = "vibbs.dod@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
package-mode = false
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-tenants (>=3.8.0,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "boto3 (>=1.39.14,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
